refactoring: merge security header listeners into new SecurityHeadersResponseListener
- removed ResponseListener and DomainAwareSecurityHeadersSetter
