--- packages/framework/src/Component/HttpFoundation/ResponseListener.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Shopsys\FrameworkBundle\Component\HttpFoundation;
-
-use Symfony\Component\HttpKernel\Event\ResponseEvent;
-
-class ResponseListener
-{
-    public function onKernelResponse(ResponseEvent $event): void
-    {
-        $event->getResponse()->headers->set('X-Frame-Options', 'sameorigin');
-        $event->getResponse()->headers->set('X-XSS-Protection', '1; mode=block');
-        $event->getResponse()->headers->set('X-Content-Type-Options', 'nosniff');
-        $event->getResponse()->headers->set('X-Powered-By', 'Shopsys Platform');
-        $event->getResponse()->headers->set('Referrer-Policy', 'same-origin');
-    }
-}
